refactor(server): extract scrape control routes (5/8 server split)
Moves the 2 scrape-trigger endpoints to their own Blueprint:
- POST /api/scrape         (manual trigger, returns 202 with snapshot)
- GET  /api/scrape/status  (poll the in-flight/last cycle snapshot)

Both auth-gate via check_api_secret(request) — the new helper from
core/config that reads os.environ live so pytest's monkeypatch.setenv
takes effect without explicit module reload.

Subtlety: the route imports the `scrape_loop` MODULE (not the function
inside it) and calls scrape_loop.run_scrape_async(...). This is a
deliberate idiom — monkeypatch.setattr(core.scrape_loop,
"run_scrape_async", stub) needs the route to look the function up live
on each call. Binding `from core.scrape_loop import run_scrape_async`
captures the function at import time, so a test patch on the source
module wouldn't reach us.

Test fix: test_concurrent_scrape_requests was patching the legacy
`server._run_scrape_async` alias. After this split the live call site is
core.scrape_loop.run_scrape_async, so the test now patches both
locations (defense in depth). Existing assertion + flow unchanged.

server.py shrinks 721 → 675 LOC (-46).

Tests: 219 passed (unchanged).
